♻️ Refactor: 파워유저 계산후 저장기능 추가

+++ src/main/java/com/codeit/duckhu/domain/user/dto/PowerUserStatsDto.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 public record PowerUserStatsDto(
         UUID userId,
         Double reviewScoreSum,
-        Integer likedCount,
-        Integer commentCount
+        Long likedCount,
+        Long commentCount
 ) {
 }
